requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
DateTime==4.3
decorator==5.1.0
distlib==0.3.3
dnspython==2.1.0
email-validator==1.1.3
filelock==3.3.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
infinity==1.5
intervals==0.9.2
is-disposable-email==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
node==0.9.28
odict==1.8.0
passlib==1.7.4
pipenv==2021.11.5.post0
pipfile==0.0.2
platformdirs==2.4.0
plumber==1.6
protobuf==3.19.1
psycopg2-binary==2.9.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
validators==0.18.2
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
WTForms==2.3.3
WTForms-Components==0.10.5
wtforms-validators==1.0.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
